nigeria changes

+++ apps/health_campaign_field_worker_app/lib/blocs/auth/auth.dart
@@ -14,6 +14,7 @@ import '../../models/entities/individual.dart';
 import '../../models/entities/roles_type.dart';
 import '../../models/role_actions/role_actions_model.dart';
 import '../../utils/environment_config.dart';
+import '../../utils/utils.dart';
 
 // part 'auth.freezed.dart' need to be added to auto generate the files for freezed model
 part 'auth.freezed.dart';
@@ -139,13 +140,26 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
   //_onLogout event logs out the user and deletes the saved user details from local storage
   FutureOr<void> _onLogout(AuthLogoutEvent event, AuthEmitter emit) async {
     try {
-      emit(const AuthLoadingState());
-      await localSecureStore.deleteAll();
-      await localSecureStore.setBoundaryRefetch(true);
+      final isConnected = await getIsConnected();
+      if (isConnected) {
+        final accessToken = await localSecureStore.accessToken;
+        final user = await localSecureStore.userRequestModel;
+        final tenantId = user?.tenantId;
+        await authRepository.logOutUser(
+          logoutPath: Constants.logoutUserPath,
+          queryParameters: {
+            'tenantId': tenantId.toString(),
+          },
+          body: {'access_token': accessToken},
+        );
+        await localSecureStore.deleteAll();
+
+        emit(const AuthUnauthenticatedState());
+      }
     } catch (error) {
-      rethrow;
+      await localSecureStore.deleteAll();
+      emit(const AuthUnauthenticatedState());
     }
-    emit(const AuthUnauthenticatedState());
   }
 }
